web_attendance/backend/requirements.txt
# Face Recognition Attendance - Backend (Windows)
# Python 3.10 ~ 3.12 권장 (3.13은 TensorFlow 미지원)
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
python-multipart>=0.0.17
pillow>=10.0.0
numpy>=1.24.0,<2.0.0
opencv-python-headless>=4.8.0
tensorflow>=2.15.0,<2.18.0
aiosqlite>=0.20.0
pydantic-settings>=2.6.0
openpyxl>=3.1.0
